% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirTableSchema.R
\name{base_schema}
\alias{base_schema}
\title{Base schema}
\usage{
base_schema(BaseId = NULL, api_key = NULL)
}
\arguments{
\item{BaseId}{Base identifier}

\item{api_key}{API key}
}
\value{

}
\description{
Returns the schema of the tables in the specified base
(https://airtable.com/api/meta).
}
