% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirTableSchema.R
\name{get_schema}
\alias{get_schema}
\title{Get AirTable Schema}
\usage{
get_schema(api_key = NULL, names = NULL)
}
\arguments{
\item{api_key}{AirTable API Key. If not specified, tries to read from
`.Renviron` (e.g., `Sys.getenv("AIRTABLE_API_KEY")`)}

\item{names}{Names of bases to retrieve schema for.}
}
\value{
A nested tibble with columns `BaseID`, `name`, `permissionLevel`  and
`schema`.
}
\description{
Retrieves AirTable schema available to an API key. Bases can be selected
using the optinal `names` argument.
}
