% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirTableSchema.R
\name{list_bases}
\alias{list_bases}
\title{List bases}
\usage{
list_bases(api_key = NULL)
}
\arguments{
\item{api_key}{API key}
}
\value{

}
\description{
Returns the list of bases the API key can access in the order they appear on
the user's home screen. The result will be truncated to only include the
first 1000 bases (https://airtable.com/api/meta)
}
